correct something
