+++ platformio.ini
@@ -11,13 +11,13 @@
 [env] ; preparation for different environments
 platform = espressif8266
 framework = arduino
-custom_prog_version = 2.5.1
+custom_prog_version = 2.5.0
 
 build_flags = 
 			-DVERSION=${this.custom_prog_version}
 			-DPIO_SRC_NAM="Daly2MQTT"
-extra_scripts = pre:tools/pre_compile.py
-				;pre:tools/minify_html.py
+extra_scripts = pre:tools/mini_html.py
+				pre:tools/pre_compile.py
 				post:tools/post_compile.py
 
 lib_deps =
