+++ src/html.h
@@ -2,29 +2,679 @@
 
 #include <Arduino.h>  // PROGMEM
 
-const char HTML_CONFIRM_RESET[] PROGMEM = R"rawliteral(%HEAD_TEMPLATE% <figure class=text-center><h1>Erase all Data?</h1></figure><div class="d-grid gap-2"><a class="btn btn-danger"href=/reset role=button>Yes</a><a class="btn btn-primary"href=/settings role=button>No</a></div> %FOOT_TEMPLATE%)rawliteral";
+const char HTML_CONFIRM_RESET[] PROGMEM = R"rawliteral(%HEAD_TEMPLATE%
 
-const char HTML_FOOT[] PROGMEM = R"rawliteral(<figure class=text-center>Daly2MQTT <a id=software_version>%SOFTWARE_VERSION%</a> By <a href=https://github.com/softwarecrash/Daly2MQTT target=_blank>Softwarecrash</a><a href=http://creativecommons.org/licenses/by-nc-sa/4.0/ rel=license target=_blank><img alt="Creative Commons License"src=https://licensebuttons.net/l/by-nc-sa/4.0/80x15.png style=border-width:0></a></figure><div style="display: none;"id=update_alert><figure class=text-center><a id=fwdownload target=_blank>Download the latest version <b id=gitversion></b></a></figure></div><script>$(document).ready(function(){$.getJSON("https://api.github.com/repos/softwarecrash/Daly2MQTT/releases/latest",function(){}).done(function(a){console.log("get data from github done success");$('#fwdownload').attr('href',a.html_url);$('#gitversion').text(a.tag_name.substring(1));let b=a.tag_name.substring(1).split('.').map(e=>parseInt(e));let c="%SWVERSION%".split('.').map(e=>parseInt(e));let d="";for(i=0;i<b.length;i++){if(b[i]===c[i]){d+="e"}else if(b[i]>c[i]){d+="m"}else{d+="l"}};if(!d.match(/[l|m]/g)){console.log("Git-Version equal, nothing to do.")}else if(d.split('e').join('')[0]=="m"){console.log("Git-Version higher, activate notification.");document.getElementById("update_alert").style.display=''}else{console.log("Git-Version lower, nothing to do.")}}).fail(function(){console.log("error can not get version")})})</script>)rawliteral";
+<figure class="text-center"><h1>Erase all Data?</h1></figure>
+<div class="d-grid gap-2">
+<a class="btn btn-danger" href="/reset" role="button">Yes</a>
+<a class="btn btn-primary" href="/settings" role="button">No</a>
+</div>
 
-const char HTML_HEAD[] PROGMEM = R"rawliteral(<!doctypehtml><html lang=en xml:lang=en><meta content=text/html;charset=UTF-8 http-equiv=content-type><meta content=width=device-width,initial-scale=1 name=viewport><link rel="shortcut icon"href=data:image/x-icon;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8YQUAAAAJcEhZcwAADsQAAA7EAZUrDhsAAAHsSURBVDhPrZLLSxtRFMZHY9A2Wou0FB+oiZUWTaPEVo15qInBjamSqLFVk0rV0ke0saAmgg90o5v42NQHCmYkXehC/wbFuK//gG6yyiIki0Dg69zDEEgp46L+4HKHc8/3zTnnXg7/yf0b7O7tIZVK0XcikaBdCjKIx+MI8jzGxj+hWadHu9mC94NDlHAXZBAIrENTVw+O49LrlaYOfv8cJUnBnV9cYNjlJlGOPBfZMnnaZGrqBw6DQTH131AFLtcHEjDxQ0UBunvsMBhMWFhcoiQpyODj6Fj6rz6fH03NOmpL2/BGmEkLnZdXVGJ+YRFPnj5DKPSLxAzu5uYWkUgEXba3KHxchP39AyhVVVheXoFarYHZ0oHPX76iplYNj2cCptY29DsHRLlYQTKZxOnpGaLRKFRV1dAbjLAJhsUlZdBqX8PpfIeS0jKhCiW+e73gsmQkZpABIxaL4ff1NbXxzeNBVnYOFPmPUP3iJazWTuQ9UNCQJyYnKef4+IR0aQMGzx/Bbu/Fz+0dDA4Nw+0ewfTMLFZX1+imWIw9NEdvPzY2t0iTYeAT7r1Fb4TBaBIeUwfNRal6TiU7HH0UY2dt7WY0NulIk2HwN2y4FosV8tw8XIbDYjQTSYPw1RUC6xvU0mGQF6OZSBrcDfAHIwsaPAvZdQgAAAAASUVORK5CYII=><link crossorigin href=https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css integrity=sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3 rel=stylesheet><link href=https://cdn.jsdelivr.net/npm/bootstrap-icons@1.3.0/font/bootstrap-icons.css rel=stylesheet><script src=https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js></script><script crossorigin integrity=sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p src=https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js></script><script src=" https://cdn.jsdelivr.net/npm/chart.js@4.3.0/dist/chart.umd.min.js "></script><title>DALY BMS 2 MQTT</title><body><noscript><strong>We're sorry but it doesn't work properly without JavaScript enabled. Please enable it to continue.</strong></noscript><div class="container-md col-md-4">)rawliteral";
+%FOOT_TEMPLATE%
+<p hidden>Hidden Helper</p> )rawliteral";
+const char HTML_FOOT[] PROGMEM = R"rawliteral(</br>
+<figure class="text-center">
+    Daly2MQTT <a id="software_version">%SOFTWARE_VERSION%</a> By <a href="https://github.com/softwarecrash/Daly2MQTT"
+        target="_blank">Softwarecrash</a>
+    <a rel="license" href="http://creativecommons.org/licenses/by-nc-sa/4.0/" target="_blank"><img
+            alt="Creative Commons License" style="border-width:0"
+            src="https://licensebuttons.net/l/by-nc-sa/4.0/80x15.png" /></a>
+</figure>
+</div>
+<div id="update_alert" style="display: none;">
+    <figure class="text-center"><a id="fwdownload" target="_blank">Download the latest version <b
+                id="gitversion"></b></a></figure>
+</div>
 
-const char HTML_MAIN[] PROGMEM = R"rawliteral(%HEAD_TEMPLATE% <figure class=text-center><h2 id=devicename></h2></figure><div class="row gx-0 mb-1"style="padding-bottom: 0rem;padding-top: 0px; display: none;"id=alert><div style="text-align: center;"id=alert role=alert><i id=alert_icon></i><span id=alert_message></span></div></div><div class="row gx-0 mb-2"><div class=col id=ClickSOC><div class=progress style=height:1.8rem;><div aria-valuemax=100 aria-valuemin=0 aria-valuenow=0 class=progress-bar id=packSOC role=progressbar style=width:0%;height:1.8rem;></div></div></div></div><div class="row gx-0 mb-2"style="display: none;"id=cellRow><div class="col card chart-container"style="position: relative; height:20vh; width:80vw"><canvas id=chart></canvas></div></div><div class="row gx-0 mb-2"><div class=col><div class=bg-light>Package:</div></div><div class=col><div class=bg-light><span id=packV></span><span id=packA></span><span id=packP></span></div></div></div><div class="row gx-0 mb-2"><div class=col><div class=bg-light>Remaining Capacity:</div></div><div class=col><div class=bg-light><span id=packRes></span></div></div></div><div class="row gx-0 mb-2"><div class=col><div class=bg-light>Charge Cycles:</div></div><div class=col><div class=bg-light><span id=packCycles></span></div></div></div><div class="row gx-0 mb-2"><div class=col><div class=bg-light>Temperature:</div></div><div class=col><div class=bg-light><span id=packTemp></span></div></div></div><div class="row gx-0 mb-2"><div class=col><div class=bg-light>Cell Difference:</div></div><div class=col><div class=bg-light><span id=cellDiff></span></div></div></div><div class="row gx-0 mb-2"><div class=col><div class=bg-light>Cells ↑/↓:</div></div><div class=col><div class=bg-light><span id=cellH></span><span id=cellL></span></div></div></div><div class="row gx-0 mb-2"><div class=col><div class=bg-light>BMS Status:</div></div><div class=col><div class=bg-light><span id=status></span><button class="btn btn-warning"style="padding: 0px;font-size: 12px;"id=wakebms type=button>Wake BMS</button></div></div></div><div class="row gx-0 mb-2"><div class=col><div class=bg-light>Charge MOS:</div></div><div class=col><div class="bg-light form-check form-switch"><input class=form-check-input id=chargeFetState role=switch type=checkbox></div></div></div><div class="row gx-0 mb-2"><div class=col><div class=bg-light>Discharge MOS:</div></div><div class=col><div class="bg-light form-check form-switch"><input class=form-check-input id=disChargeFetState role=switch type=checkbox></div></div></div><div class="row gx-0 mb-2"><div class=col><div class=bg-light>Balance State:</div></div><div class=col><div class="bg-light form-check form-switch"><input class=form-check-input disabled id=cellBalanceActive role=switch type=checkbox></div></div></div><div class="row gx-0 mb-2"style=%ESP01%><div class=col><div class=bg-light>Relais Output:</div></div><div class=col><div class="bg-light form-check form-switch"><input class=form-check-input disabled id=relaisOutputActive role=switch type=checkbox></div></div></div><div class="d-grid gap-2"><a class="btn btn-primary btn-block"href=/settings role=button>Settings</a></div><script>$(document).ready(function(){initWebSocket();initButton();setInterval(refreshAlert,5000)});var gateway=`ws://${window.location.host}/ws`;var websocket;var ctx;var cellChart;var createBarChart=true;var alertListArr=[];var alertListitem=0;var kickRefresh=true;function initWebSocket(){console.log('Trying to open a WebSocket connection...');websocket=new WebSocket(gateway);websocket.onopen=onOpen;websocket.onclose=onClose;websocket.onerror=onError;websocket.onmessage=onMessage}function onOpen(a){console.log('Connection opened')}function onClose(a){document.getElementById("status").innerHTML='WS Closed';console.log('Connection closed');setTimeout(initWebSocket,3000)}function onError(a){document.getElementById("status").innerHTML='WS Lost';console.log('Connection lost')}function onMessage(a){var b=JSON.parse(a.data);document.getElementById("devicename").innerHTML=b.Device.Name;document.getElementById("packV").innerHTML=b.Pack.Voltage+ 'V ';document.getElementById("packA").innerHTML=b.Pack.Current+ 'A  ';document.getElementById("packP").innerHTML=Math.round(b.Pack.Power)+ 'W  ';document.getElementById("packSOC").innerHTML=b.Pack.SOC+ '%%';$('#packSOC').width(b.Pack.SOC+ "%").attr('aria-valuenow',b.Pack.SOC);document.getElementById("packRes").innerHTML=b.Pack.Remaining_mAh+ 'mAh ';document.getElementById("cellH").innerHTML=b.Pack.High_CellNr+ '↑'+ b.Pack.High_CellV+ 'V ';document.getElementById("cellL").innerHTML=b.Pack.Low_CellNr+ '↓'+ b.Pack.Low_CellV+ 'V ';document.getElementById("packCycles").innerHTML=b.Pack.Cycles+ ' ';document.getElementById("packTemp").innerHTML=b.Pack.BMS_Temp+ '°C ';document.getElementById("cellDiff").innerHTML=b.Pack.Cell_Diff+ 'mV ';document.getElementById("status").innerHTML=b.Pack.Status;document.getElementById("chargeFetState").checked=b.Pack.ChargeFET;document.getElementById("disChargeFetState").checked=b.Pack.DischargeFET;document.getElementById("cellBalanceActive").checked=b.Pack.Balance_Active;document.getElementById("relaisOutputActive").checked=b.Device.Relais_Active;BarChart(b);alert(b);if(b.Pack.Status=="offline"){document.getElementById("status").style.color="red";document.getElementById("wakebms").style.display=''}else{document.getElementById("status").style.color="black";document.getElementById("wakebms").style.display='none'};if(b.Device.Relais_Manual){relaisOutputActive.removeAttribute("disabled")}else{relaisOutputActive.setAttribute('disabled','disabled')}}function alert(a){alertListArr=[];if(a.Device.ESP_VCC<2.6){alertListArr.push("ESP Volt to Low")};if(!a.Pack.Fail_Codes.length==0){var b;var c=a.Pack.Fail_Codes.split(',');for(var d in c){b=c[d];alertListArr.push(c[d])}};if(alertListArr.length==0){document.getElementById("alert").style.display='none'}else{document.getElementById("alert").style.display=''};if(kickRefresh){refreshAlert();kickRefresh=false}}function refreshAlert(){var a;if(alertListitem<alertListArr.length- 1){a=alertListArr[alertListitem];alertListitem++}else{a=alertListArr[alertListitem];alertListitem=0};if(a[a.length- 1]=="1"){document.getElementById("alert_icon").className="bi bi-info-circle-fill";document.getElementById("alert").className="row gx-0 mb-2 alert alert-info"}else if(a[a.length- 1]=="2"){document.getElementById("alert_icon").className="bi bi-exclamation-circle-fill";document.getElementById("alert").className="row gx-0 mb-2 alert alert-warning"}else{document.getElementById("alert_icon").className="bi bi-x-circle-fill";document.getElementById("alert").className="row gx-0 mb-2 alert alert-danger"};document.getElementById('alert_message').innerHTML=a}function initButton(){document.getElementById('chargeFetState').addEventListener('click',ChargeFetSwitch);document.getElementById('disChargeFetState').addEventListener('click',DischargeFetSwitch);document.getElementById('relaisOutputActive').addEventListener('click',RelaisOutputSwitch);document.getElementById('wakebms').addEventListener('click',wakeBms);document.getElementById('ClickSOC').addEventListener('click',cellState)}function wakeBms(){let a;a='wake_bms';websocket.send(a)}function cellState(){var a=document.getElementById("cellRow");if(a.style.display==="none"){a.style.display=""}else{a.style.display="none"}}function ChargeFetSwitch(){let a;if(document.getElementById('chargeFetState').checked){a='chargeFetSwitch_on'}else{a='chargeFetSwitch_off'};websocket.send(a)}function RelaisOutputSwitch(){let a;if(document.getElementById('relaisOutputActive').checked){a='relaisOutputSwitch_on'}else{a='relaisOutputSwitch_off'};websocket.send(a)}function DischargeFetSwitch(){let b;if(document.getElementById('disChargeFetState').checked){b='dischargeFetSwitch_on';websocket.send(b)}else{b='dischargeFetSwitch_off';var a=confirm('Are you sure to disable the DISCHARGE MOS?! You maybe create your own personal blackout!');if(a){websocket.send(b)}else{document.getElementById("disChargeFetState").checked=true}}}function BarChart(a){var b=Object.values(a.CellV);var c=[];var d=[];var e=[];var f=[];var g=0;var h=0;for(let i=0;i<b.length;i++){if(i%2==0){c.push(b[i]);e[g]=g+ 1;if(g==a.Pack.High_CellNr- 1){f[g]='DarkBlue'}else if(g==a.Pack.Low_CellNr- 1){f[g]='LightSkyBlue'}else{f[g]='#0a58ca'};if(g==a.Pack.High_CellNr- 1){f[g]='DarkBlue'};g=g+ 1}else{d.push(b[i]);if(b[i]==true){f[h]='BlueViolet'};h=h+ 1}};if(createBarChart==true){createBarChart=false;ctx=document.getElementById("chart").getContext('2d');cellChart=new Chart(ctx,{type:'bar',data:{labels:e,datasets:[{label:'Cell Voltage',backgroundColor:f,borderColor:'rgb(13, 110, 253)',data:c}]},options:{maintainAspectRatio:false,plugins:{legend:{display:false},title:{display:false},label:{display:false}},scales:{y:{min:a.Pack.cell_lVt,max:a.Pack.cell_hVt},x:{display:true}}}})}else{cellChart.data.datasets.pop();cellChart.data.datasets.push({backgroundColor:f,borderColor:'rgb(10, 88, 202)',labels:e,label:'Cell Voltage',data:c});cellChart.update('none')}}</script> %FOOT_TEMPLATE%)rawliteral";
+<script>
+    $(document).ready(function () {
+        $.getJSON("https://api.github.com/repos/softwarecrash/Daly2MQTT/releases/latest", function () {
+        })
+            .done(function (data) {
+                console.log("get data from github done success");
+                $('#fwdownload').attr('href', data.html_url);
+                $('#gitversion').text(data.tag_name.substring(1));
+                let x = data.tag_name.substring(1).split('.').map(e => parseInt(e));
+                let y = "%SWVERSION%".split('.').map(e => parseInt(e));
+                let z = "";
+                for (i = 0; i < x.length; i++) { if (x[i] === y[i]) { z += "e"; } else if (x[i] > y[i]) { z += "m"; } else { z += "l"; } }
+                if (!z.match(/[l|m]/g)) {
+                    console.log("Git-Version equal, nothing to do.");
+                } else if (z.split('e').join('')[0] == "m") {
+                    console.log("Git-Version higher, activate notification.");
+                    document.getElementById("update_alert").style.display = '';
+                } else { console.log("Git-Version lower, nothing to do."); }
+            })
+            .fail(function () {
+                console.log("error can not get version");
+            });
+    });
+</script>
 
-const char HTML_REBOOT[] PROGMEM = R"rawliteral(%HEAD_TEMPLATE% <figure class=text-center><h1>Rebooting</h1><h2 id=wait>.</h2></figure><div class="d-grid gap-2"><a class="btn btn-primary"href=/ role=button>Main</a></div><script>$(document).ready(function(){window.dotsGoingUp=true;var a=window.setInterval(function(){var d=document.getElementById("wait");if(window.dotsGoingUp)d.innerHTML+=".";else{d.innerHTML=d.innerHTML.substring(1,d.innerHTML.length);if(d.innerHTML===".")window.dotsGoingUp=true};if(d.innerHTML.length>9)window.dotsGoingUp=false},100);setTimeout(b,2000);function b(){setInterval(c,1000)}function c(){$.get("/").done(function(){window.location="/";console.log("success")}).fail(function(){console.log("failed.")})}})</script> %FOOT_TEMPLATE%)rawliteral";
+</body>
 
-const char HTML_SETTINGS[] PROGMEM = R"rawliteral(%HEAD_TEMPLATE% <figure class=text-center><h1>Settings</h1></figure><div class="d-grid gap-2"><form action=/update enctype=multipart/form-data method=POST><div style="display: none;"class=input-group id=flash_block><input aria-describedby=inputGroupFileAddon04 aria-label=Upload class=form-control id=inputGroupFile04 name=update type=file><input class="btn btn-outline-secondary"id=inputGroupFileAddon04 type=submit value=Update></div></form><div class="row gx-0 mb-2"style="display: none;"id=flash_alert><div class="alert alert-warning"style="text-align: center;"><span><b>ESP FLASH TO SMALL FOR OTA-UPDATE</b></span></div></div><a class="btn btn-primary"href=/settingsedit role=button>Configure</a><a class="btn btn-warning"href=/reboot role=button>Reboot</a><a class="btn btn-primary"href=/confirmreset role=button>Reset ESP</a><a class="btn btn-primary"href=/ role=button>Back</a></div> %FOOT_TEMPLATE% <script>var Flash_Size = %Flash_Size%;
-        console.log(Flash_Size);
-        if (Flash_Size < 1048575) {
-            document.getElementById("flash_alert").style.display = '';
-            document.getElementById("flash_block").style.display = 'none';
-        }else{
-            document.getElementById("flash_alert").style.display = 'none';
-            document.getElementById("flash_block").style.display = '';
-        }</script>)rawliteral";
+</html>)rawliteral";
+const char HTML_HEAD[] PROGMEM = R"rawliteral(<!DOCTYPE html>
+<html lang="en" xml:lang="en">
 
-const char HTML_SETTINGS_EDIT[] PROGMEM = R"rawliteral(%HEAD_TEMPLATE% <figure class=text-center><h1>Edit Configuration</h1></figure><form action=/settingssave enctype=multipart/form-data method=POST><div class="input-group mb-3"><span class="input-group-text w-50"id=devicenamedesc>Device Name</span><input aria-describedby=devicenamedesc class=form-control id=devicename maxlength=35 name=post_deviceName value=%DEVICE_NAME%></div><div class="input-group mb-2"><span class="input-group-text w-50"id=mqttserverdesc>MQTT Server</span><input aria-describedby=mqttserverdesc class=form-control id=mqttserver maxlength=35 name=post_mqttServer value=%MQTT_SERVER%></div><div class="input-group mb-2"><span class="input-group-text w-50"id=mqttportdesc>MQTT Port</span><input aria-describedby=mqttportdesc class=form-control id=mqttport maxlength=5 name=post_mqttPort value=%MQTT_PORT%></div><div class="input-group mb-2"><span class="input-group-text w-50"id=mqttuserdesc>MQTT User</span><input aria-describedby=mqttuserdesc class=form-control id=mqttuser maxlength=35 name=post_mqttUser value=%MQTT_USER%></div><div class="input-group mb-2"><span class="input-group-text w-50"id=mqttpassworddesc>MQTT Password</span><input aria-describedby=mqttpassworddesc class=form-control id=mqttpassword maxlength=35 name=post_mqttPassword type=password value=%MQTT_PASS%></div><div class="input-group mb-2"><span class="input-group-text w-50"id=mqtttopicdesc>MQTT Topic</span><input aria-describedby=mqtttopicdesc class=form-control id=mqtttopic maxlength=35 name=post_mqttTopic value=%MQTT_TOPIC%></div><div class="input-group mb-2"><span class="input-group-text w-50"id=mqttrefreshdesc>MQTT Refresh (sec)</span><input aria-describedby=mqttrefreshdesc class=form-control id=mqttrefresh maxlength=5 name=post_mqttRefresh value=%MQTT_REFRESH%></div><div class="input-group mb-3"><span class="input-group-text w-50"id=mqttjsondesc>MQTT Json Style</span><div class="form-switch form-control mqtt-settings-switch"style="width:50%%; text-align: center;"><input class="form-check-input form control"%mqtt_json% aria-describedby=mqttjsondesc id=mqttjson name=post_mqttjson role=switch type=checkbox value=true></div></div><div class="row gx-0 mb-2"id=esp01_settings style=%ESP01%><div class="input-group mb-2"><span class="input-group-text w-100"><b>BMS-Wakeup Settings</b></span></div><div class="input-group mb-2"><span class="input-group-text w-50"id=wakeupenabledesc>Enable BMS Wakeup GPIO%WAKEUP_PIN%</span><div class="form-switch form-control mqtt-settings-switch"style="width:50%%; text-align: center;"><input class="form-check-input form control"%bms_wake% aria-describedby=wakeupenabledesc id=wakeupenable name=post_wakeupenable role=switch type=checkbox value=true></div></div><div class="input-group mb-2"><span class="input-group-text w-100"><b>Output Settings</b></span></div><div class="input-group mb-2"><span class="input-group-text w-50"id=relaisenabledesc>Enable Output on GPIO%RELAISPIN%</span><div class="form-switch form-control mqtt-settings-switch"style="width:50%%; text-align: center;"><input class="form-check-input form control"%relais_enable% aria-describedby=relaisenabledesc id=relaisenable name=post_relaisenable role=switch type=checkbox value=true></div></div><div class="input-group mb-2"><span class="input-group-text w-50"id=relaisinvertdesc>Invert GPIO Output</span><div class="form-switch form-control mqtt-settings-switch"style="width:50%%; text-align: center;"><input class="form-check-input form control"%relais_invert% aria-describedby=relaisinvertdesc id=relaisinvert name=post_relaisinvert role=switch type=checkbox value=true></div></div><div class="input-group mb-2"><span class="input-group-text w-50"id=relaisfailsafedesc>Failsafe Mode</span><div class="form-switch form-control mqtt-settings-switch"style="width:50%%; text-align: center;"><input class="form-check-input form control"%relais_failsave% aria-describedby=relaisfailsafedesc id=relaisfailsafe name=post_relaisfailsafe role=switch type=checkbox value=true></div></div><div class="input-group mb-2"><span class="input-group-text w-50"id=relaisfunctiondesc>Function</span><select aria-describedby=relaisfunctiondesc class=form-select id=relaisfunction name=post_relaisfunction><option value=0>Lowest Cell Voltage<option value=1>Highest Cell Voltage<option value=2>Pack Cell Voltage<option value=3>Temperature<option value=4>Manual over Web or MQTT</select></div><div class="input-group mb-2"><span class="input-group-text w-50"id=relaiscomparsiondesc>Comparsion</span><select aria-describedby=relaiscomparsiondesc class=form-select id=relaiscomparsion name=post_relaiscomparsion><option value=0>Higher or equal than<option value=1>Lower or equal than</select></div><div class="input-group mb-2"><span class="input-group-text w-50"id=relaissetvaluedesc>Value</span><input aria-describedby=relaissetvaluedesc class=form-control id=relaissetvalue max=100 min=-100 name=post_relaissetvalue step=0.001 type=number value=%RELAIS_VALUE%></div><div class="input-group mb-2"><span class="input-group-text w-50"id=relaishysteresisdesc>Hysteresis</span><input aria-describedby=relaishysteresisdesc class=form-control id=relaishysteresis max=100 min=-100 name=post_relaishysteresis step=0.001 type=number value=%RELAIS_HYST%></div></div><div class="d-grid gap-2"><input class="btn btn-primary"value="Save settings"type=submit><a class="btn btn-primary"href=/settings role=button>Back</a></div></form><script>$(document).ready(function (load) {
-        $("#relaisfunction").val(%RELAIS_FUNCTION%);
-        $("#relaiscomparsion").val(%RELAIS_COMP%);
+<head>
+    <meta http-equiv="content-type" content="text/html;charset=UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="shortcut icon"
+        href="data:image/x-icon;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8YQUAAAAJcEhZcwAADsQAAA7EAZUrDhsAAAHsSURBVDhPrZLLSxtRFMZHY9A2Wou0FB+oiZUWTaPEVo15qInBjamSqLFVk0rV0ke0saAmgg90o5v42NQHCmYkXehC/wbFuK//gG6yyiIki0Dg69zDEEgp46L+4HKHc8/3zTnnXg7/yf0b7O7tIZVK0XcikaBdCjKIx+MI8jzGxj+hWadHu9mC94NDlHAXZBAIrENTVw+O49LrlaYOfv8cJUnBnV9cYNjlJlGOPBfZMnnaZGrqBw6DQTH131AFLtcHEjDxQ0UBunvsMBhMWFhcoiQpyODj6Fj6rz6fH03NOmpL2/BGmEkLnZdXVGJ+YRFPnj5DKPSLxAzu5uYWkUgEXba3KHxchP39AyhVVVheXoFarYHZ0oHPX76iplYNj2cCptY29DsHRLlYQTKZxOnpGaLRKFRV1dAbjLAJhsUlZdBqX8PpfIeS0jKhCiW+e73gsmQkZpABIxaL4ff1NbXxzeNBVnYOFPmPUP3iJazWTuQ9UNCQJyYnKef4+IR0aQMGzx/Bbu/Fz+0dDA4Nw+0ewfTMLFZX1+imWIw9NEdvPzY2t0iTYeAT7r1Fb4TBaBIeUwfNRal6TiU7HH0UY2dt7WY0NulIk2HwN2y4FosV8tw8XIbDYjQTSYPw1RUC6xvU0mGQF6OZSBrcDfAHIwsaPAvZdQgAAAAASUVORK5CYII=">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.3.0/font/bootstrap-icons.css" rel="stylesheet">
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
+        crossorigin="anonymous"></script>
+
+    <script src=" https://cdn.jsdelivr.net/npm/chart.js@4.3.0/dist/chart.umd.min.js "></script>
+
+    <title>DALY BMS 2 MQTT</title>
+</head>
+
+<body>
+    <noscript>
+        <strong>We're sorry but it doesn't work properly without JavaScript enabled. Please enable it to
+            continue.</strong>
+    </noscript>
+    <div class="container-md col-md-4">)rawliteral";
+const char HTML_MAIN[] PROGMEM = R"rawliteral(%HEAD_TEMPLATE%
+
+<figure class="text-center">
+    <h2 id="devicename"></h2>
+</figure>
+
+<div class="row gx-0 mb-1" id="alert" style="padding-bottom: 0rem;padding-top: 0px; display: none;">
+    <div id="alert" class="" role="alert" style="text-align: center;">
+        <i id="alert_icon" class=""></i>
+        <span id="alert_message"></span>
+    </div>
+</div>
+
+<div class="row gx-0 mb-2">
+    <div class="col" id="ClickSOC">
+        <div class="progress" style="height:1.8rem;">
+            <div id="packSOC" class="progress-bar" role="progressbar" style="width:0%;height:1.8rem;" aria-valuenow="0"
+                aria-valuemin="0" aria-valuemax="100"></div>
+        </div>
+    </div>
+</div>
+
+<div id="cellRow" class="row gx-0 mb-2" style="display: none;">
+    <div class="col card chart-container" style="position: relative; height:20vh; width:80vw">
+        <canvas id="chart"></canvas>
+    </div>
+</div>
+
+<div class="row gx-0 mb-2">
+    <div class="col">
+        <div class="bg-light">Package:</div>
+    </div>
+    <div class="col">
+        <div class="bg-light"><span id="packV"></span><span id="packA"></span><span id="packP"></span></div>
+    </div>
+</div>
+
+<div class="row gx-0 mb-2">
+    <div class="col">
+        <div class="bg-light">Remaining Capacity: </div>
+    </div>
+    <div class="col">
+        <div class="bg-light"><span id="packRes"></span></div>
+    </div>
+</div>
+
+<div class="row gx-0 mb-2">
+    <div class="col">
+        <div class="bg-light">Charge Cycles: </div>
+    </div>
+    <div class="col">
+        <div class="bg-light"><span id="packCycles"></span></div>
+    </div>
+</div>
+
+<div class="row gx-0 mb-2">
+    <div class="col">
+        <div class="bg-light">Temperature: </div>
+    </div>
+    <div class="col">
+        <div class="bg-light"><span id="packTemp"></span></div>
+    </div>
+</div>
+
+<div class="row gx-0 mb-2">
+    <div class="col">
+        <div class="bg-light">Cell Difference: </div>
+    </div>
+    <div class="col">
+        <div class="bg-light"><span id="cellDiff"></span></div>
+    </div>
+</div>
+
+<div class="row gx-0 mb-2">
+    <div class="col">
+        <div class="bg-light">Cells ↑/↓:</div>
+    </div>
+    <div class="col">
+        <div class="bg-light"><span id="cellH"></span><span id="cellL"></span></div>
+    </div>
+</div>
+
+<div class="row gx-0 mb-2">
+    <div class="col">
+        <div class="bg-light">BMS Status: </div>
+    </div>
+    <div class="col">
+
+        <div class="bg-light">
+            <span id="status"></span>
+            <button id="wakebms" type="button" class="btn btn-warning" style="padding: 0px;font-size: 12px;">Wake
+                BMS</button>
+        </div>
+    </div>
+</div>
+
+<div class="row gx-0 mb-2">
+    <div class="col">
+        <div class="bg-light">Charge MOS:</div>
+    </div>
+    <div class="col">
+        <div class="bg-light form-check form-switch">
+            <input class="form-check-input" type="checkbox" role="switch" id="chargeFetState">
+        </div>
+    </div>
+</div>
+
+<div class="row gx-0 mb-2">
+    <div class="col">
+        <div class="bg-light">Discharge MOS:</div>
+    </div>
+    <div class="col">
+        <div class="bg-light form-check form-switch"><input class="form-check-input" type="checkbox" role="switch"
+                id="disChargeFetState"></div>
+    </div>
+</div>
+
+<div class="row gx-0 mb-2">
+    <div class="col">
+        <div class="bg-light">Balance State: </div>
+    </div>
+    <div class="col">
+        <div class="bg-light form-check form-switch"><input class="form-check-input" type="checkbox" role="switch"
+                id="cellBalanceActive" disabled></div>
+    </div>
+</div>
+
+<div class="row gx-0 mb-2" style="%ESP01%">
+    <div class="col">
+        <div class="bg-light">Relais Output: </div>
+    </div>
+    <div class="col">
+        <div class="bg-light form-check form-switch"><input class="form-check-input" type="checkbox" role="switch"
+                id="relaisOutputActive" disabled></div>
+    </div>
+</div>
+
+<div class="d-grid gap-2">
+    <a class="btn btn-primary btn-block" href="/settings" role="button">Settings</a>
+</div>
+
+
+<script>
+    $(document).ready(function () {
+        initWebSocket();
+        initButton();
+        setInterval(refreshAlert, 5000);
+    });
+    var gateway = `ws://${window.location.host}/ws`;
+    var websocket;
+    var ctx;
+    var cellChart;
+    var createBarChart = true;
+    var alertListArr = [];
+    var alertListitem = 0;
+    var kickRefresh = true;
+
+    function initWebSocket() {
+        console.log('Trying to open a WebSocket connection...');
+        websocket = new WebSocket(gateway);
+        websocket.onopen = onOpen;
+        websocket.onclose = onClose;
+        websocket.onerror = onError;
+        websocket.onmessage = onMessage;
+    }
+    function onOpen(event) {
+        console.log('Connection opened');
+    }
+    function onClose(event) {
+        document.getElementById("status").innerHTML = 'WS Closed';
+        console.log('Connection closed');
+        setTimeout(initWebSocket, 3000);
+    }
+    function onError(event) {
+        document.getElementById("status").innerHTML = 'WS Lost';
+        console.log('Connection lost');
+    }
+    function onMessage(event) {
+        var data = JSON.parse(event.data);
+        document.getElementById("devicename").innerHTML = data.Device.Name;
+        document.getElementById("packV").innerHTML = data.Pack.Voltage + 'V ';
+        document.getElementById("packA").innerHTML = data.Pack.Current + 'A  ';
+        document.getElementById("packP").innerHTML = Math.round(data.Pack.Power) + 'W  ';
+        document.getElementById("packSOC").innerHTML = data.Pack.SOC + '%%';
+        $('#packSOC').width(data.Pack.SOC + "%").attr('aria-valuenow', data.Pack.SOC);
+        document.getElementById("packRes").innerHTML = data.Pack.Remaining_mAh + 'mAh ';
+
+        document.getElementById("cellH").innerHTML = data.Pack.High_CellNr + '↑' + data.Pack.High_CellV + 'V ';
+        document.getElementById("cellL").innerHTML = data.Pack.Low_CellNr + '↓' + data.Pack.Low_CellV + 'V ';
+
+        document.getElementById("packCycles").innerHTML = data.Pack.Cycles + ' ';
+        document.getElementById("packTemp").innerHTML = data.Pack.BMS_Temp + '°C ';
+        document.getElementById("cellDiff").innerHTML = data.Pack.Cell_Diff + 'mV ';
+        document.getElementById("status").innerHTML = data.Pack.Status;
+        document.getElementById("chargeFetState").checked = data.Pack.ChargeFET;
+        document.getElementById("disChargeFetState").checked = data.Pack.DischargeFET;
+        document.getElementById("cellBalanceActive").checked = data.Pack.Balance_Active;
+        document.getElementById("relaisOutputActive").checked = data.Device.Relais_Active;
+
+        BarChart(data);
+        alert(data);
+
+        if (data.Pack.Status == "offline") {
+            document.getElementById("status").style.color = "red";
+            document.getElementById("wakebms").style.display = '';
+        } else {
+            document.getElementById("status").style.color = "black";
+            document.getElementById("wakebms").style.display = 'none';
+        }
+        if (data.Device.Relais_Manual) {
+            relaisOutputActive.removeAttribute("disabled")
+        } else {
+            relaisOutputActive.setAttribute('disabled', 'disabled');
+        }
+    }
+
+    function alert(data) {
+        alertListArr = [];
+        if (data.Device.ESP_VCC < 2.6) {
+            alertListArr.push("ESP Volt to Low");
+        }
+        if (!data.Pack.Fail_Codes.length == 0) {
+            var i;
+            var i_list = data.Pack.Fail_Codes.split(',');
+            for (var i_index in i_list) {
+                i = i_list[i_index];
+                alertListArr.push(i_list[i_index]);
+            }
+        }
+        if (alertListArr.length == 0) {
+            document.getElementById("alert").style.display = 'none';
+        } else {
+            document.getElementById("alert").style.display = '';
+        }
+        if (kickRefresh) {
+            refreshAlert();
+            kickRefresh = false;
+        }
+    }
+
+    function refreshAlert() {
+        var alertValue;
+        if (alertListitem < alertListArr.length - 1) {
+            alertValue = (alertListArr[alertListitem]);
+            alertListitem++;
+        } else {
+            alertValue = (alertListArr[alertListitem]);
+            alertListitem = 0;
+        }
+        if (typeof alertValue !== 'undefined') {
+            if (alertValue[alertValue.length - 1] == "1") {
+
+                document.getElementById("alert_icon").className = "bi bi-info-circle-fill";
+                document.getElementById("alert").className = "row gx-0 mb-2 alert alert-info";
+            } else if (alertValue[alertValue.length - 1] == "2") {
+                document.getElementById("alert_icon").className = "bi bi-exclamation-circle-fill";
+                document.getElementById("alert").className = "row gx-0 mb-2 alert alert-warning";
+            } else {
+                document.getElementById("alert_icon").className = "bi bi-x-circle-fill";
+                document.getElementById("alert").className = "row gx-0 mb-2 alert alert-danger";
+            }
+        }
+        document.getElementById('alert_message').innerHTML = (alertValue);
+    }
+
+    function initButton() {
+        document.getElementById('chargeFetState').addEventListener('click', ChargeFetSwitch);
+        document.getElementById('disChargeFetState').addEventListener('click', DischargeFetSwitch);
+        document.getElementById('relaisOutputActive').addEventListener('click', RelaisOutputSwitch);
+        document.getElementById('wakebms').addEventListener('click', wakeBms);
+        document.getElementById('ClickSOC').addEventListener('click', cellState);
+    }
+
+    function wakeBms() {
+        let switchVal;
+        switchVal = 'wake_bms';
+        websocket.send(switchVal);
+    }
+
+    function cellState() {
+        var x = document.getElementById("cellRow");
+        if (x.style.display === "none") {
+            x.style.display = "";
+        } else {
+            x.style.display = "none";
+        }
+    }
+
+    function ChargeFetSwitch() {
+        let switchVal;
+        if (document.getElementById('chargeFetState').checked) { switchVal = 'chargeFetSwitch_on' }
+        else { switchVal = 'chargeFetSwitch_off' }
+        websocket.send(switchVal);
+    }
+
+    function RelaisOutputSwitch() {
+        let switchVal;
+        if (document.getElementById('relaisOutputActive').checked) { switchVal = 'relaisOutputSwitch_on' }
+        else { switchVal = 'relaisOutputSwitch_off' }
+        websocket.send(switchVal);
+    }
+
+    function DischargeFetSwitch() {
+        let switchVal;
+        if (document.getElementById('disChargeFetState').checked) {
+            switchVal = 'dischargeFetSwitch_on';
+            websocket.send(switchVal);
+        }
+        else {
+            switchVal = 'dischargeFetSwitch_off';
+            var check = confirm('Are you sure to disable the DISCHARGE MOS?! You maybe create your own personal blackout!');
+            if (check) {
+                websocket.send(switchVal);
+            } else {
+                document.getElementById("disChargeFetState").checked = true;
+            }
+        }
+    }
+
+    function BarChart(dataObj) {
+        var tmpCellV = Object.values(dataObj.CellV);
+        var cellVoltages = [];
+        var cellBalances = [];
+        var cellCount = [];
+        var cellColor = [];
+        var tmpCountV = 0;
+        var tmpCountB = 0;
+        for (let i = 0; i < tmpCellV.length; i++) {
+            if (i % 2 == 0) {
+                cellVoltages.push(tmpCellV[i]);
+                cellCount[tmpCountV] = tmpCountV + 1;
+                if (tmpCountV == dataObj.Pack.High_CellNr - 1) { cellColor[tmpCountV] = 'DarkBlue'; }
+                else if (tmpCountV == dataObj.Pack.Low_CellNr - 1) { cellColor[tmpCountV] = 'LightSkyBlue'; }
+                else { cellColor[tmpCountV] = '#0a58ca'; }
+                if (tmpCountV == dataObj.Pack.High_CellNr - 1) { cellColor[tmpCountV] = 'DarkBlue'; }
+                tmpCountV = tmpCountV + 1;
+            } else {
+                cellBalances.push(tmpCellV[i]);
+                if (tmpCellV[i] == true) { cellColor[tmpCountB] = 'BlueViolet'; }
+                tmpCountB = tmpCountB + 1;
+            }
+        }
+        if (createBarChart == true) {
+            createBarChart = false;
+            ctx = document.getElementById("chart").getContext('2d');
+            cellChart = new Chart(ctx, {
+                type: 'bar',
+                data: {
+                    labels: cellCount,
+                    datasets: [{
+                        label: 'Cell Voltage',
+                        backgroundColor: cellColor,
+                        borderColor: 'rgb(13, 110, 253)',
+                        data: cellVoltages,
+                    }]
+                },
+                options: {
+                    maintainAspectRatio: false,
+                    plugins: {
+                        legend: { display: false },
+                        title: { display: false },
+                        label: { display: false }
+                    },
+                    scales: {
+                        y: {
+                            min: dataObj.Pack.cell_lVt,
+                            max: dataObj.Pack.cell_hVt
+                        },
+                        x: {
+                            display: true
+                        }
+                    }
+                },
+            });
+        } else {
+            cellChart.data.datasets.pop();
+            cellChart.data.datasets.push({
+                backgroundColor: cellColor,
+                borderColor: 'rgb(10, 88, 202)',
+                labels: cellCount,
+                label: 'Cell Voltage',
+                data: cellVoltages
+            });
+            cellChart.update('none');
+        }
+    }
+</script>
+
+%FOOT_TEMPLATE%
+<p hidden>Hidden Helper</p> )rawliteral";
+const char HTML_REBOOT[] PROGMEM = R"rawliteral(%HEAD_TEMPLATE%
+
+<figure class="text-center"><h1>Rebooting </h1><h2 id="wait">.</h2></figure>
+<div class="d-grid gap-2">
+<a class="btn btn-primary" href="/" role="button">Main</a>
+</div>
+
+<script>
+$(document).ready(function () {
+    window.dotsGoingUp = true;
+    var dots = window.setInterval( function() {
+        var wait = document.getElementById("wait");
+        if ( window.dotsGoingUp ) 
+            wait.innerHTML += ".";
+        else {
+            wait.innerHTML = wait.innerHTML.substring(1, wait.innerHTML.length);
+            if ( wait.innerHTML === ".")
+                window.dotsGoingUp = true;
+        }
+        if ( wait.innerHTML.length > 9 )
+            window.dotsGoingUp = false;
+        }, 100);
+
+setTimeout(startCheck, 2000);
+
+function startCheck(){
+  setInterval(serverReachable, 1000);
+}
+
+function serverReachable() {
+$.get("/").done(function () {
+  window.location = "/";
+  console.log("success");
+}).fail(function () {
+   console.log("failed.");
+});
+}
+
+});
+</script>
+
+%FOOT_TEMPLATE%
+<p hidden>Hidden Helper</p> )rawliteral";
+const char HTML_SETTINGS[] PROGMEM = R"rawliteral(%HEAD_TEMPLATE%
+
+<figure class="text-center">
+    <h1>Settings</h1>
+</figure>
+<div class="d-grid gap-2">
+    <form method="POST" action="/update" enctype="multipart/form-data">
+        <div class="input-group" id="flash_block" style="display: none;">
+            <input class="form-control" id="inputGroupFile04" aria-describedby="inputGroupFileAddon04"
+                aria-label="Upload" type="file" name="update"><input class="btn btn-outline-secondary"
+                id="inputGroupFileAddon04" type="submit" value="Update">
+        </div>
+    </form>
+    <div class="row gx-0 mb-2" id="flash_alert" style="display: none;">
+        <div class="alert alert-warning" style="text-align: center;">
+            <span><b>ESP FLASH TO SMALL FOR OTA-UPDATE</b></span>
+        </div>
+    </div>
+    <a class="btn btn-primary" href="/settingsedit" role="button">Configure</a>
+    <a class="btn btn-warning" href="/reboot" role="button">Reboot</a>
+    <a class="btn btn-primary" href="/confirmreset" role="button">Reset ESP</a>
+    <a class="btn btn-primary" href="/" role="button">Back</a>
+</div>
+
+<script>
+    var Flash_Size = Number("%Flash_Size%");
+    console.log(Flash_Size);
+    if (Flash_Size < 1048575) {
+        document.getElementById("flash_alert").style.display = '';
+        document.getElementById("flash_block").style.display = 'none';
+    } else {
+        document.getElementById("flash_alert").style.display = 'none';
+        document.getElementById("flash_block").style.display = '';
+    }
+
+</script>
+
+%FOOT_TEMPLATE%
+<p hidden>Hidden Helper</p> )rawliteral";
+const char HTML_SETTINGS_EDIT[] PROGMEM = R"rawliteral(%HEAD_TEMPLATE%
+
+<figure class="text-center">
+    <h1>Edit Configuration</h1>
+</figure>
+<form method="POST" action="/settingssave" enctype="multipart/form-data">
+    <div class="input-group mb-3">
+        <span class="input-group-text w-50" id="devicenamedesc">Device Name</span>
+        <input type="text" class="form-control" aria-describedby="devicenamedesc" id="devicename" name="post_deviceName" maxlength="35"
+            value="%DEVICE_NAME%">
+    </div>
+    <div class="input-group mb-2">
+        <span class="input-group-text w-50" id="mqttserverdesc">MQTT Server</span>
+        <input type="text" class="form-control" aria-describedby="mqttserverdesc" id="mqttserver" name="post_mqttServer" maxlength="35"
+            value="%MQTT_SERVER%">
+    </div>
+    <div class="input-group mb-2">
+        <span class="input-group-text w-50" id="mqttportdesc">MQTT Port</span>
+        <input type="text" class="form-control" aria-describedby="mqttportdesc" id="mqttport" name="post_mqttPort" maxlength="5"
+            value="%MQTT_PORT%">
+    </div>
+    <div class="input-group mb-2">
+        <span class="input-group-text w-50" id="mqttuserdesc">MQTT User</span>
+        <input type="text" class="form-control" aria-describedby="mqttuserdesc" id="mqttuser" name="post_mqttUser" maxlength="35"
+            value="%MQTT_USER%">
+    </div>
+    <div class="input-group mb-2">
+        <span class="input-group-text w-50" id="mqttpassworddesc">MQTT Password</span>
+        <input type="password" class="form-control" aria-describedby="mqttpassworddesc" id="mqttpassword" maxlength="35"
+            name="post_mqttPassword" value="%MQTT_PASS%">
+    </div>
+    <div class="input-group mb-2">
+        <span class="input-group-text w-50" id="mqtttopicdesc">MQTT Topic</span>
+        <input type="text" class="form-control" aria-describedby="mqtttopicdesc" id="mqtttopic" name="post_mqttTopic" maxlength="35"
+            value="%MQTT_TOPIC%">
+    </div>
+    <div class="input-group mb-2">
+        <span class="input-group-text w-50" id="mqttrefreshdesc">MQTT Refresh (sec)</span>
+        <input type="text" class="form-control" aria-describedby="mqttrefreshdesc" id="mqttrefresh" maxlength="5"
+            name="post_mqttRefresh" value="%MQTT_REFRESH%">
+    </div>
+    <div class="input-group mb-3">
+        <span class="input-group-text w-50" id="mqttjsondesc">MQTT Json Style</span>
+        <div class="form-switch form-control mqtt-settings-switch" style="width:50%%; text-align: center;">
+            <input type="checkbox" class="form-check-input form control" aria-describedby="mqttjsondesc"
+                role="switch" id="mqttjson" name="post_mqttjson" value="true" %MQTT_JSON%>
+        </div>
+    </div>
+    <div class="row gx-0 mb-2" id="esp01_settings" style="%ESP01%">
+    <div class="input-group mb-2">
+        <span class="input-group-text w-100"><b>BMS-Wakeup Settings</b></span>
+    </div>
+	<div class="input-group mb-2">
+        <span class="input-group-text w-50" id="wakeupenabledesc">Enable BMS Wakeup GPIO%WAKEUP_PIN%</span>
+		<div class="form-switch form-control mqtt-settings-switch" style="width:50%%; text-align: center;">
+            <input type="checkbox" class="form-check-input form control" aria-describedby="wakeupenabledesc"
+                role="switch" id="wakeupenable" name="post_wakeupenable" value="true" %BMS_WAKE%>
+        </div>
+    </div>
+	<div class="input-group mb-2">
+        <span class="input-group-text w-100"><b>Output Settings</b></span>
+    </div>
+	<div class="input-group mb-2">
+        <span class="input-group-text w-50" id="relaisenabledesc">Enable Output on GPIO%RELAISPIN%</span>
+		<div class="form-switch form-control mqtt-settings-switch" style="width:50%%; text-align: center;">
+            <input type="checkbox" class="form-check-input form control" aria-describedby="relaisenabledesc"
+                role="switch" id="relaisenable" name="post_relaisenable" value="true" %RELAIS_ENABLE%>
+        </div>
+    </div>
+	<div class="input-group mb-2">
+        <span class="input-group-text w-50" id="relaisinvertdesc">Invert GPIO Output</span>
+		<div class="form-switch form-control mqtt-settings-switch" style="width:50%%; text-align: center;">
+            <input type="checkbox" class="form-check-input form control" aria-describedby="relaisinvertdesc"
+                role="switch" id="relaisinvert" name="post_relaisinvert" value="true" %RELAIS_INVERT%>
+        </div>
+    </div>
+
+    <div class="input-group mb-2">
+        <span class="input-group-text w-50" id="relaisfailsafedesc">Failsafe Mode</span>
+		<div class="form-switch form-control mqtt-settings-switch" style="width:50%%; text-align: center;">
+            <input type="checkbox" class="form-check-input form control" aria-describedby="relaisfailsafedesc"
+                role="switch" id="relaisfailsafe" name="post_relaisfailsafe" value="true" %RELAIS_FAILSAVE%>
+        </div>
+    </div>
+
+	<div class="input-group mb-2">
+        <span class="input-group-text w-50" id="relaisfunctiondesc">Function</span>
+        <select class="form-select" aria-describedby="relaisfunctiondesc" id="relaisfunction" name="post_relaisfunction">
+             <option value="0">Lowest Cell Voltage</option>
+			 <option value="1">Highest Cell Voltage</option>
+			 <option value="2">Pack Cell Voltage</option>
+			 <option value="3">Temperature</option>
+             <option value="4">Manual over Web or MQTT</option>
+		</select>
+    </div>
+	<div class="input-group mb-2">
+        <span class="input-group-text w-50" id="relaiscomparsiondesc">Comparsion</span>
+        <select class="form-select" aria-describedby="relaiscomparsiondesc" id="relaiscomparsion" name="post_relaiscomparsion">
+             <option value="0">Higher or equal than</option>
+			 <option value="1">Lower or equal than</option>
+		</select>
+    </div>
+	<div class="input-group mb-2">
+        <span class="input-group-text w-50" id="relaissetvaluedesc">Value</span>
+        <input type="number" step="0.001" class="form-control" aria-describedby="relaissetvaluedesc" id="relaissetvalue"
+            name="post_relaissetvalue" value="%RELAIS_VALUE%" min="-100" max="100">
+    </div>
+    <div class="input-group mb-2">
+        <span class="input-group-text w-50" id="relaishysteresisdesc">Hysteresis</span>
+        <input type="number" step="0.001" class="form-control" aria-describedby="relaishysteresisdesc" id="relaishysteresis"
+            name="post_relaishysteresis" value="%RELAIS_HYST%" min="-100" max="100">
+    </div>
+    </div>
+    <div class="d-grid gap-2">
+        <input class="btn btn-primary" type="submit" value="Save settings">
+        <a class="btn btn-primary" href="/settings" role="button">Back</a>
+    </div>
+</form>
+<script>
+    $(document).ready(function (load) {
+        $("#relaisfunction").val("%RELAIS_FUNCTION%");
+        $("#relaiscomparsion").val("%RELAIS_COMP%");
         if(document.getElementById("relaisfunction").value == 4){
             RelaisFunctionChange();
         }
@@ -42,5 +692,8 @@ const char HTML_SETTINGS_EDIT[] PROGMEM = R"rawliteral(%HEAD_TEMPLATE% <figure c
             relaissetvalue.removeAttribute("disabled");
             relaishysteresis.removeAttribute("disabled");
         }
-    }</script> %FOOT_TEMPLATE%)rawliteral";
+    }
+</script>
 
+%FOOT_TEMPLATE%
+<p hidden>Hidden Helper</p> )rawliteral";
