+++ src/webpages/HTML_CONFIRM_RESET.html
@@ -1,7 +1,10 @@
 %HEAD_TEMPLATE%
+
 <figure class="text-center"><h1>Erase all Data?</h1></figure>
 <div class="d-grid gap-2">
 <a class="btn btn-danger" href="/reset" role="button">Yes</a>
 <a class="btn btn-primary" href="/settings" role="button">No</a>
 </div>
+
 %FOOT_TEMPLATE%
+<p hidden>Hidden Helper</p>
