+++ src/webpages/HTML_HEAD.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
 <html lang="en" xml:lang="en">
+
 <head>
     <meta http-equiv="content-type" content="text/html;charset=UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
@@ -13,10 +14,11 @@
         integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
         crossorigin="anonymous"></script>
 
-<script src=" https://cdn.jsdelivr.net/npm/chart.js@4.3.0/dist/chart.umd.min.js "></script>
+    <script src=" https://cdn.jsdelivr.net/npm/chart.js@4.3.0/dist/chart.umd.min.js "></script>
 
     <title>DALY BMS 2 MQTT</title>
 </head>
+
 <body>
     <noscript>
         <strong>We're sorry but it doesn't work properly without JavaScript enabled. Please enable it to
