+++ src/webpages/HTML_REBOOT.html
@@ -40,3 +40,4 @@ $.get("/").done(function () {
 </script>
 
 %FOOT_TEMPLATE%
+<p hidden>Hidden Helper</p>
