+++ src/webpages/HTML_SETTINGS_EDIT.html
@@ -1,4 +1,5 @@
 %HEAD_TEMPLATE%
+
 <figure class="text-center">
     <h1>Edit Configuration</h1>
 </figure>
@@ -117,8 +118,8 @@
 </form>
 <script>
     $(document).ready(function (load) {
-        $("#relaisfunction").val(%RELAIS_FUNCTION%);
-        $("#relaiscomparsion").val(%RELAIS_COMP%);
+        $("#relaisfunction").val("%RELAIS_FUNCTION%");
+        $("#relaiscomparsion").val("%RELAIS_COMP%");
         if(document.getElementById("relaisfunction").value == 4){
             RelaisFunctionChange();
         }
@@ -138,4 +139,6 @@
         }
     }
 </script>
+
 %FOOT_TEMPLATE%
+<p hidden>Hidden Helper</p>
